app: Refactor explorer codebase
