+++ CodeApp/Views/InfiniteProgressView.swift
